联系人账号类型修改为unsigned int

+++ nwtClient.cpp
@@ -77,7 +77,7 @@ BOOL CnwtClientApp::InitInstance()
     int nResponse = loginDlg.DoModal();
     if (nResponse == IDOK) {
         CnwtClientDlg dlg;
-        dlg.m_own.m_account = atoll(loginDlg.m_strAccount);
+        dlg.m_own.m_account = atoi(loginDlg.m_strAccount);
         dlg.m_own.m_nickname = loginDlg.m_strNickname.GetString();
 
         m_pMainWnd = &dlg;
